nwsl/cloud-run-proxy/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.25.1
pydantic==2.5.0
google-cloud-secret-manager==2.17.0